+++ NBC_W10-11-BookCart/NBC_W10-11-BookCart/View/BookDetailView.swift
@@ -16,7 +16,6 @@ protocol BookDetailDelegate: AnyObject {
 class BookDetailView: UIView {
     
     weak var delegate: BookDetailDelegate?
-    
     var bookInfo: Document? {
         didSet {
             titleLabel.text = bookInfo?.title
