+++ NBC_W10-11-BookCart/NBC_W10-11-BookCart/View/Component/TableCell.swift
@@ -46,7 +46,7 @@ class TableCell: UITableViewCell {
         fatalError("init(coder:) has not been implemented")
     }
     
-    // MARK: - SetUp
+    // MARK: - Setup
     private func setup() {
         [bookTitleLabel, authorLabel, priceLabel]
             .forEach { addSubview($0) }
