+++ NBC_W10-11-BookCart/NBC_W10-11-BookCart/ViewController/BookDetailVC.swift
@@ -60,15 +60,7 @@ extension BookDetailVC: SearchBookVCDelegate {
         searchBookDelegate?.didSelectBook()
     }
 }
-extension UIViewController {
-    /// 간단한 경고창을 생성하는 함수
-    func showAlert(title: String = "경고", message: String) {
-        let alert = UIAlertController(title: title, message: message, preferredStyle: .alert)
-        let okAction = UIAlertAction(title: "확인", style: .default, handler: nil)
-        alert.addAction(okAction)
-        self.present(alert, animated: true, completion: nil)
-    }
-}
+
 
 @available(iOS 17.0, *)
 #Preview {
